Moved more logic in the name builders
