+++ Service/IPlanetNamesBuilder.cs
@@ -1,0 +1,11 @@
+using StellarisNameListGenerator.Models;
+
+namespace StellarisNameListGenerator.Service
+{
+    public interface IPlanetNamesBuilder
+    {
+        string Build(NameList nameList);
+
+        string GetRandomName(NameList nameList);
+    }
+}
