+++ Service/IShipNamesBuilder.cs
@@ -1,0 +1,11 @@
+using StellarisNameListGenerator.Models;
+
+namespace StellarisNameListGenerator.Service
+{
+    public interface IShipNamesBuilder
+    {
+        string Build(NameList nameList);
+
+        string GetRandomName(NameList nameList);
+    }
+}
